+++ src/main/java/com/falsepattern/zigbrains/zon/completion/ZonCompletionContributor.java
@@ -1,0 +1,86 @@
+/*
+ * Copyright 2023 FalsePattern
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.falsepattern.zigbrains.zon.completion;
+
+import com.falsepattern.zigbrains.zon.parser.ZonFile;
+import com.falsepattern.zigbrains.zon.psi.ZonProperty;
+import com.falsepattern.zigbrains.zon.psi.ZonPropertyPlaceholder;
+import com.falsepattern.zigbrains.zon.psi.ZonStruct;
+import com.falsepattern.zigbrains.zon.psi.ZonTypes;
+import com.falsepattern.zigbrains.zon.psi.impl.ZonPsiImplUtil;
+import com.intellij.codeInsight.completion.CompletionContributor;
+import com.intellij.codeInsight.completion.CompletionParameters;
+import com.intellij.codeInsight.completion.CompletionProvider;
+import com.intellij.codeInsight.completion.CompletionResultSet;
+import com.intellij.codeInsight.completion.CompletionType;
+import com.intellij.codeInsight.lookup.LookupElementBuilder;
+import com.intellij.patterns.PlatformPatterns;
+import com.intellij.util.ProcessingContext;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.List;
+import java.util.Set;
+
+public class ZonCompletionContributor extends CompletionContributor {
+    private static final List<String> ZON_ROOT_KEYS = List.of("name", "version", "dependencies");
+    private static final List<String> ZON_DEP_KEYS = List.of("url", "hash");
+    public ZonCompletionContributor() {
+        extend(CompletionType.BASIC,
+               PlatformPatterns.psiElement()
+                               .withParent(PlatformPatterns.psiElement(ZonTypes.PROPERTY_PLACEHOLDER))
+                               .withSuperParent(3, PlatformPatterns.psiElement(ZonFile.class)),
+               new CompletionProvider<>() {
+                    @Override
+                    protected void addCompletions(@NotNull CompletionParameters parameters, @NotNull ProcessingContext context, @NotNull CompletionResultSet result) {
+                        var placeholder = (ZonPropertyPlaceholder) parameters.getPosition().getParent();
+                        var zonStruct = (ZonStruct) placeholder.getParent();
+                        var keys = ZonPsiImplUtil.getKeys(zonStruct);
+                        doAddCompletions(placeholder.getText().startsWith("."), keys, ZON_ROOT_KEYS, result);
+                    }
+                });
+        extend(CompletionType.BASIC,
+               PlatformPatterns.psiElement()
+                               .withParent(PlatformPatterns.psiElement(ZonTypes.PROPERTY_PLACEHOLDER))
+                               .withSuperParent(3, PlatformPatterns.psiElement(ZonTypes.VALUE))
+                               .withSuperParent(6, PlatformPatterns.psiElement(ZonTypes.VALUE))
+                               .withSuperParent(9, PlatformPatterns.psiElement(ZonFile.class)),
+               new CompletionProvider<>() {
+                   @Override
+                   protected void addCompletions(@NotNull CompletionParameters parameters, @NotNull ProcessingContext context, @NotNull CompletionResultSet result) {
+                       var placeholder = (ZonPropertyPlaceholder) parameters.getPosition().getParent();
+                       var depStruct = (ZonStruct) placeholder.getParent();
+                       var parentProperty = (ZonProperty) depStruct.getParent()
+                                                                   .getParent()
+                                                                   .getParent()
+                                                                   .getParent()
+                                                                   .getParent();
+                       if (!"dependencies".equals(ZonPsiImplUtil.getText(parentProperty.getIdentifier()))) {
+                           return;
+                       }
+                       doAddCompletions(placeholder.getText().startsWith("."), ZonPsiImplUtil.getKeys(depStruct), ZON_DEP_KEYS, result);
+                   }
+               });
+    }
+    private static void doAddCompletions(boolean hasDot, Set<String> current, List<String> expected, CompletionResultSet result) {
+        for (var key: expected) {
+            if (current.contains(key)) {
+                continue;
+            }
+            result.addElement(LookupElementBuilder.create(hasDot ? key : "." + key));
+        }
+    }
+}