+++ CHANGELOG.md
@@ -24,7 +24,7 @@ Changelog structure reference:
 - Basic parser and PSI tree
 - Basic syntax highlighting
 - Color settings page
-- Brace matching
+- Brace and quote matching
 - Code completion
 - Indentation